chore: fix filter type

+++ graphql/documents.ts
@@ -13,7 +13,7 @@ export const searchPlayers = gql`
 `
 
 export const playerProfile = gql`
-  query PlayerProfile($player: BigInt!, $year: Decimal = ${dayjs().year()}) {
+  query PlayerProfile($player: BigInt!, $year: Int = ${dayjs().year()}) {
     player(id: $player) {
       id
       name
